requirements.txt
# Core dependencies
tensorflow>=2.10.0,<2.14.0
tensorflow-addons>=0.19.0
numpy>=1.21.0,<1.24.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pillow>=8.3.0
opencv-python-headless>=4.5.0  # Headless version to avoid Qt issues

# Data processing
pyyaml>=5.4.0
h5py>=3.1.0

# Utilities
tqdm>=4.62.0
psutil>=5.8.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# Additional visualization
plotly>=5.3.0

# Development tools (optional)
black>=21.7b0
flake8>=3.9.0
pytest>=6.2.0
pytest-cov>=2.12.0